requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asyncache==0.3.1
bcrypt==4.1.2
cachetools==5.3.3
cffi==1.16.0
click==8.1.7
content-size-limit-asgi==0.1.5
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
pendulum==3.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
time-machine==2.14.1
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
